code/requirements.txt
tqdm
dicom2nifti
scikit-image>=0.14
medpy
scipy
batchgenerators>=0.23
scikit-learn
SimpleITK
pandas
requests
nibabel
tifffile
matplotlib
adabelief-pytorch
torchinfo
openpyxl
einops


notebook
tensorboard
monai
pytorch-ignite

natsort
medcam